# packages/contracts/openapi/control-api.yaml
openapi: 3.0.3
info:
  title: KofClub Control API
  version: 0.1.0
  description: |
    Control-plane HTTP API (accounts, clubs, roles, table configuration,
    virtual-chip ledger administration, hand history).

    Conventions (see docs/adr/ADR-009-wire-conventions.md):
    - All endpoints are versioned under `/v1` (health probes are unversioned).
    - JSON bodies use camelCase; enums are UPPERCASE strings.
    - Identifiers are UUID strings (UUIDv7 preferred).
    - Timestamps are UTC RFC 3339 strings.
    - Chip amounts are integers (int64). Floating point is never used.
    - Every response carries an `X-Request-Id` header; clients may send one.
    - State-changing requests that may be retried accept an `Idempotency-Key` header.
    - Errors always use the `ErrorEnvelope` shape with a machine-readable `code`.

    Virtual chips have no monetary value. This API intentionally contains no
    deposit, withdrawal, cash-out or payment functionality (ADR-006).
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /v1/clubs/{clubId}/tables:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    post:
      tags: [tables]
      operationId: createTable
      summary: Create a No-Limit Hold'em table (ADMIN+)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTableRequest'
      responses:
        '201':
          description: Table created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableDetail'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    get:
      tags: [tables]
      operationId: listClubTables
      summary: Tables of a club
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableList'
  /v1/tables/{tableId}:
    parameters:
      - $ref: '#/components/parameters/TableId'
    get:
      tags: [tables]
      operationId: getTable
      summary: Table configuration and seated players
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableDetail'
        '404':
          $ref: '#/components/responses/Error'
  /v1/tables/{tableId}/seat:
    parameters:
      - $ref: '#/components/parameters/TableId'
    post:
      tags: [tables]
      operationId: takeSeat
      summary: Buy in from the club wallet and take a seat
      description: The Idempotency-Key (if given) makes retried buy-ins safe.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatRequest'
      responses:
        '200':
          description: Seated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatResult'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /v1/tables/{tableId}/leave:
    parameters:
      - $ref: '#/components/parameters/TableId'
    post:
      tags: [tables]
      operationId: leaveTable
      summary: Leave the table; the stack is cashed out to the club wallet (after the current hand if needed)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Left or leaving after the hand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveResult'
        '409':
          $ref: '#/components/responses/Error'
  /v1/tables/{tableId}/close:
    parameters:
      - $ref: '#/components/parameters/TableId'
    post:
      tags: [tables]
      operationId: closeTable
      summary: Close the table for good (ADMIN+); seated players are cashed out to their wallets
      description: |
        No new hands are dealt. When no hand is in progress every seat is
        cashed out immediately (status CLOSED); otherwise after the current
        hand (status CLOSING). Safe to retry.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Closure applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloseTableResult'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/tables/{tableId}/state:
    parameters:
      - $ref: '#/components/parameters/TableId'
    get:
      tags: [tables]
      operationId: getTableState
      summary: Viewer-sanitized table snapshot (realtime clients use TABLE_SNAPSHOT instead)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Snapshot
          content:
            application/json:
              schema:
                $ref: './realtime.yaml#/components/schemas/TableSnapshot'
        '403':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TableId:
      name: tableId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    ClubId:
      name: clubId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Unique key making a retried state-changing request a no-op.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    Error:
      description: Machine-readable error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp
    ChipAmount:
      type: integer
      format: int64
      description: Integer amount of virtual chips (no monetary value).
...
    Table:
      type: object
      required:
        [
          id,
          clubId,
          name,
          gameType,
          maxSeats,
          smallBlind,
          bigBlind,
          buyInMin,
          buyInMax,
          actionTimeoutSec,
          status,
          createdBy,
          createdAt,
          seatedCount,
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        clubId:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        gameType:
          type: string
          enum: [NLHE]
        maxSeats:
          type: integer
        smallBlind:
          $ref: '#/components/schemas/ChipAmount'
        bigBlind:
          $ref: '#/components/schemas/ChipAmount'
        buyInMin:
          $ref: '#/components/schemas/ChipAmount'
        buyInMax:
          $ref: '#/components/schemas/ChipAmount'
        actionTimeoutSec:
          type: integer
        status:
          type: string
          enum: [OPEN, CLOSED]
        createdBy:
          $ref: '#/components/schemas/Uuid'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        seatedCount:
          type: integer
    TableSeat:
      type: object
      additionalProperties: false
      required: [seatNo, userId, username, stack, sittingOut]
      properties:
        seatNo:
          type: integer
        userId:
          $ref: '#/components/schemas/Uuid'
        username:
          type: string
        stack:
          $ref: '#/components/schemas/ChipAmount'
        sittingOut:
          type: boolean
    TableDetail:
      allOf:
        - $ref: '#/components/schemas/Table'
        - type: object
          required: [seats]
          properties:
            seats:
              type: array
              items:
                $ref: '#/components/schemas/TableSeat'
    TableList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Table'
    CreateTableRequest:
      type: object
      additionalProperties: false
      required: [name, smallBlind, bigBlind, buyInMin, buyInMax]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 64
        maxSeats:
          type: integer
          minimum: 2
          maximum: 10
          default: 6
        smallBlind:
          $ref: '#/components/schemas/ChipAmount'
        bigBlind:
          $ref: '#/components/schemas/ChipAmount'
        buyInMin:
          $ref: '#/components/schemas/ChipAmount'
        buyInMax:
          $ref: '#/components/schemas/ChipAmount'
        actionTimeoutSec:
          type: integer
          minimum: 5
          maximum: 120
          default: 20
    SeatRequest:
      type: object
      additionalProperties: false
      required: [buyIn]
      properties:
        seatNo:
          type: integer
          minimum: 1
          maximum: 10
          description: Omit to take the first free seat.
        buyIn:
          $ref: '#/components/schemas/ChipAmount'
    SeatResult:
      type: object
      additionalProperties: false
      required: [tableId, seatNo, stack, seq]
      properties:
        tableId:
          $ref: '#/components/schemas/Uuid'
        seatNo:
          type: integer
        stack:
          $ref: '#/components/schemas/ChipAmount'
        seq:
          type: integer
          format: int64
    CloseTableResult:
      type: object
      additionalProperties: false
      required: [tableId, status, seated]
      properties:
        tableId:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum: [CLOSED, CLOSING]
          description: CLOSING while a hand finishes; seats are then cashed out.
        seated:
          type: integer
          description: Players still seated (0 once CLOSED).
    LeaveResult:
      type: object
      additionalProperties: false
      required: [tableId, status, cashOut]
      properties:
        tableId:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum: [LEFT, LEAVING_AFTER_HAND]
        cashOut:
          $ref: '#/components/schemas/ChipAmount'
...
    ErrorCode:
      type: string
      description: |
        Stable machine-readable error codes shared by HTTP and realtime APIs.
        Clients must branch on `code`, never on `message`.
      enum:
        # generic
        - VALIDATION_FAILED
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL
        - SERVICE_UNAVAILABLE
        - IDEMPOTENCY_CONFLICT
        # auth
        - AUTH_REQUIRED
        - AUTH_INVALID_CREDENTIALS
        - AUTH_TOKEN_INVALID
        - AUTH_TOKEN_EXPIRED
        - AUTH_REFRESH_INVALID
        - AUTH_REFRESH_REUSED
        - AUTH_SESSION_REVOKED
        - ACCOUNT_SUSPENDED
        - EMAIL_TAKEN
        - USERNAME_TAKEN
        - FORBIDDEN
        # clubs
        - CLUB_NOT_FOUND
        - NOT_CLUB_MEMBER
        - CLUB_BANNED
        - ALREADY_CLUB_MEMBER
        - INVITE_INVALID
        - ROLE_CHANGE_NOT_ALLOWED
        # ledger
        - INSUFFICIENT_CHIPS
        - LEDGER_INVARIANT_VIOLATION
        # tables / game
        - TABLE_NOT_FOUND
        - TABLE_CLOSED
        - TABLE_FULL
        - SEAT_TAKEN
        - ALREADY_SEATED
        - PLAYER_NOT_SEATED
        - INVALID_BUY_IN
        - NOT_YOUR_TURN
        - ILLEGAL_ACTION
        - INVALID_RAISE
        - HAND_NOT_ACTIVE
        - STALE_GAME_STATE
        - ACTION_ALREADY_PROCESSED
        - TABLE_UNAVAILABLE
        # history
        - HAND_NOT_FOUND
        # tournaments (phase 3)
        - TOURNAMENT_NOT_OPEN
    ErrorBody:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable summary (not for programmatic use).
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'